backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["yibrino <yibrahmehari128@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django-rest-framework = "^0.1.0"
django-cors-headers = "^4.3.1"
environ = "^1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
